// src/app/model/spot.ts
export class Spot {
    name: String | undefined;
	latitude: Number | undefined;
	longitude: Number | undefined;
	// phone_number: String | undefined;
	address: String | undefined;
	images: String[] | undefined;
    open_hour: String | undefined;
    close_hour: String | undefined;
    active: boolean | undefined;
	radius: Number | undefined;
}